Support caching PartProcessor outputs via `CachedPartProcessor`.
This change adds a new Processor type that provides basic cache-aside caching. It is a wrapper around PartProcessor (regular stream-based Processors are not supported, only per-part ones) and a pluggable cache backends.
A cache key is computed for each incoming `Part`. A processor key is prepended to it and we llok up result in the cache. If present. result is returned. Otherwise the wrapped processor is called and its output is written to the cache.

PiperOrigin-RevId: 784131778