Adds web.py, which includes a `FetchRequestExtractor` to extract URLs from text content and yield `FetchRequest` parts
This is a groundwork for implementing a processor that fetches pages and documents from the Web. This PR does not include that functionality yet.

We want the logic for detecting URLs in the stream to be optional. It is very convenient close to the UI, e.g. when user types prompts. But applying it universally may lead to fetching content accidentally.

PiperOrigin-RevId: 786658266